rename specific subjects to state
