+++ libs/observable-api/index.ts
@@ -1,7 +1,7 @@
 export * from './observer.controller';
 export * from './observer';
-export * from './unique-behavior-subject';
-export * from './unique-array-behavior-subject';
-export * from './unique-object-behavior-subject';
+export * from './deep-state';
+export * from './array-state';
+export * from './object-state';
 export * from './create-observable-part.method'
 export * from './append-to-frozen-array.method'
