+++ libs/observable-api/object-state.ts
@@ -1,15 +1,15 @@
-import { UniqueBehaviorSubject } from "./unique-behavior-subject";
+import { DeepState } from "./deep-state";
 
 /**
  * @export
- * @class UniqueObjectBehaviorSubject
- * @extends {UniqueBehaviorSubject<T>}
- * @description - A RxJS UniqueObjectBehaviorSubject which deepFreezes the object-data to ensure its not manipulated from any implementations.
+ * @class ObjectState
+ * @extends {DeepState<T>}
+ * @description - A RxJS BehaviorSubject which deepFreezes the object-data to ensure its not manipulated from any implementations.
  * Additionally the Subject ensures the data is unique, not updating any Observes unless there is an actual change of the content.
  *
- * The UniqueObjectBehaviorSubject provides methods to append data when the data is an Object.
+ * The ObjectState provides methods to append data when the data is an Object.
  */
-export class UniqueObjectBehaviorSubject<T> extends UniqueBehaviorSubject<T> {
+export class ObjectState<T> extends DeepState<T> {
 
 	/**
 	 * @method append
@@ -17,7 +17,7 @@ export class UniqueObjectBehaviorSubject<T> extends UniqueBehaviorSubject<T> {
 	 * @description - Append some new data to this Object.
 	 * @example <caption>Example append some data.</caption>
 	 * const data = {key: 'myKey', value: 'myInitialValue'};
-	 * const mySubject = new UniqueObjectBehaviorSubject(data)
+	 * const mySubject = new ObjectState(data)
 	 * mySubject.append({value: 'myNewValue'})
 	 */
 	update(partialData: Partial<T>) {
