+++ src/backoffice/documents/document-blueprints/document-blueprint.tree.store.ts
@@ -1,7 +1,7 @@
 import { DocumentBlueprintResource, DocumentTreeItem } from '@umbraco-cms/backend-api';
 import { tryExecuteAndNotify } from '@umbraco-cms/resources';
 import { UmbContextToken } from '@umbraco-cms/context-api';
-import { createObservablePart, UniqueArrayBehaviorSubject } from '@umbraco-cms/observable-api';
+import { createObservablePart, ArrayState } from '@umbraco-cms/observable-api';
 import { UmbStoreBase } from '@umbraco-cms/store';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
@@ -18,7 +18,7 @@ export const UMB_DocumentBlueprint_TREE_STORE_CONTEXT_TOKEN = new UmbContextToke
 export class UmbDocumentBlueprintTreeStore extends UmbStoreBase {
 
 
-	#data = new UniqueArrayBehaviorSubject<DocumentTreeItem>([], (x) => x.key);
+	#data = new ArrayState<DocumentTreeItem>([], (x) => x.key);
 
 
 	constructor(host: UmbControllerHostInterface) {
