+++ src/backoffice/media/media-types/media-type.detail.store.ts
@@ -1,6 +1,6 @@
 import type { DataTypeDetails } from '@umbraco-cms/models';
 import { UmbContextToken } from '@umbraco-cms/context-api';
-import { UniqueArrayBehaviorSubject } from '@umbraco-cms/observable-api';
+import { ArrayState } from '@umbraco-cms/observable-api';
 import { UmbStoreBase } from '@umbraco-cms/store';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
@@ -17,7 +17,7 @@ export const UMB_MEDIA_TYPE_DETAIL_STORE_CONTEXT_TOKEN = new UmbContextToken<Umb
 export class UmbMediaTypeDetailStore extends UmbStoreBase {
 
 
-	private _data = new UniqueArrayBehaviorSubject<DataTypeDetails>([], (x) => x.key);
+	private _data = new ArrayState<DataTypeDetails>([], (x) => x.key);
 
 
 	constructor(host: UmbControllerHostInterface) {
