+++ src/backoffice/media/media/media.tree.store.ts
@@ -2,7 +2,7 @@ import type { Observable } from 'rxjs';
 import { MediaResource, ContentTreeItem } from '@umbraco-cms/backend-api';
 import { tryExecuteAndNotify } from '@umbraco-cms/resources';
 import { UmbContextToken } from '@umbraco-cms/context-api';
-import { createObservablePart, UniqueArrayBehaviorSubject } from '@umbraco-cms/observable-api';
+import { createObservablePart, ArrayState } from '@umbraco-cms/observable-api';
 import { UmbStoreBase } from '@umbraco-cms/store';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
@@ -22,7 +22,7 @@ export class UmbMediaTreeStore extends UmbStoreBase {
 
 
 
-	#data = new UniqueArrayBehaviorSubject<MediaTreeItem>([], (x) => x.key);
+	#data = new ArrayState<MediaTreeItem>([], (x) => x.key);
 
 
 	constructor(host: UmbControllerHostInterface) {
