+++ src/backoffice/members/member-groups/member-group.details.store.ts
@@ -1,7 +1,7 @@
 import { Observable } from 'rxjs';
 import type { MemberGroupDetails } from '@umbraco-cms/models';
 import { UmbContextToken } from '@umbraco-cms/context-api';
-import { UniqueArrayBehaviorSubject } from '@umbraco-cms/observable-api';
+import { ArrayState } from '@umbraco-cms/observable-api';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 import { UmbStoreBase } from '@umbraco-cms/store';
 
@@ -16,7 +16,7 @@ export const UMB_MEMBER_GROUP_STORE_CONTEXT_TOKEN = new UmbContextToken<UmbMembe
 export class UmbMemberGroupStore extends UmbStoreBase {
 
 
-	#groups = new UniqueArrayBehaviorSubject<MemberGroupDetails>([], x => x.key);
+	#groups = new ArrayState<MemberGroupDetails>([], x => x.key);
 	public groups = this.#groups.asObservable();
