+++ src/backoffice/members/member-types/member-type.tree.store.ts
@@ -1,7 +1,7 @@
 import { EntityTreeItem, MemberTypeResource, } from '@umbraco-cms/backend-api';
 import { tryExecuteAndNotify } from '@umbraco-cms/resources';
 import { UmbContextToken } from '@umbraco-cms/context-api';
-import { createObservablePart, UniqueArrayBehaviorSubject } from '@umbraco-cms/observable-api';
+import { createObservablePart, ArrayState } from '@umbraco-cms/observable-api';
 import { UmbStoreBase } from '@umbraco-cms/store';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 
@@ -19,7 +19,7 @@ export class UmbMemberTypeTreeStore extends UmbStoreBase {
 
 
 	// TODO: use the right type here:
-	#data = new UniqueArrayBehaviorSubject<EntityTreeItem>([], (x) => x.key);
+	#data = new ArrayState<EntityTreeItem>([], (x) => x.key);
 
 
 	constructor(host: UmbControllerHostInterface) {
