+++ src/backoffice/shared/components/workspace-property/workspace-property.context.ts
@@ -1,7 +1,7 @@
 import { UmbWorkspaceContentContext } from '../workspace/workspace-content/workspace-content.context';
 import type { DataTypeDetails } from '@umbraco-cms/models';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
-import { createObservablePart, UniqueObjectBehaviorSubject } from '@umbraco-cms/observable-api';
+import { createObservablePart, ObjectState } from '@umbraco-cms/observable-api';
 import { UmbContextConsumerController, UmbContextProviderController } from '@umbraco-cms/context-api';
 
 // If we get this from the server then we can consider using TypeScripts Partial<> around the model from the Management-API.
@@ -16,7 +16,7 @@ export type WorkspacePropertyData<ValueType> = {
 export class UmbWorkspacePropertyContext<ValueType = unknown> {
 	private _providerController: UmbContextProviderController;
 
-	private _data = new UniqueObjectBehaviorSubject<WorkspacePropertyData<ValueType>>({});
+	private _data = new ObjectState<WorkspacePropertyData<ValueType>>({});
 
 	public readonly alias = createObservablePart(this._data, (data) => data.alias);
 	public readonly label = createObservablePart(this._data, (data) => data.label);
