+++ src/backoffice/translation/dictionary/dictionary.detail.store.ts
@@ -1,6 +1,6 @@
 import type { DictionaryDetails } from '@umbraco-cms/models';
 import { UmbContextToken } from '@umbraco-cms/context-api';
-import { createObservablePart, UniqueArrayBehaviorSubject } from '@umbraco-cms/observable-api';
+import { createObservablePart, ArrayState } from '@umbraco-cms/observable-api';
 import { UmbStoreBase } from '@umbraco-cms/store';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
 import { EntityTreeItem } from '@umbraco-cms/backend-api';
@@ -19,7 +19,7 @@ export class UmbDictionaryDetailStore extends UmbStoreBase {
 
 
 	// TODO: use the right type:
-	#data = new UniqueArrayBehaviorSubject<EntityTreeItem>([], (x) => x.key);
+	#data = new ArrayState<EntityTreeItem>([], (x) => x.key);
 
 
 	constructor(host: UmbControllerHostInterface) {
