+++ src/backoffice/users/user-groups/user-group.store.ts
@@ -1,7 +1,7 @@
 import type { UserGroupDetails } from '@umbraco-cms/models';
 import { UmbContextToken } from '@umbraco-cms/context-api';
 import { UmbControllerHostInterface } from '@umbraco-cms/controller';
-import { createObservablePart, UniqueArrayBehaviorSubject } from '@umbraco-cms/observable-api';
+import { createObservablePart, ArrayState } from '@umbraco-cms/observable-api';
 import { UmbStoreBase } from '@umbraco-cms/store';
 
 // TODO: get rid of this type addition & { ... }:
@@ -18,7 +18,7 @@ export const UMB_USER_GROUP_STORE_CONTEXT_TOKEN = new UmbContextToken<UmbUserGro
 export class UmbUserGroupStore extends UmbStoreBase {
 
 
-	#groups = new UniqueArrayBehaviorSubject<UmbUserGroupStoreItemType>([], x => x.key);
+	#groups = new ArrayState<UmbUserGroupStoreItemType>([], x => x.key);
 	public groups = this.#groups.asObservable();
